# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "znskr api",
    "description": "znskr paas api for managing apps and deployments",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/buckets": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "list all buckets for the authenticated user",
        "operationId": "list_buckets",
        "responses": {
          "200": {
            "description": "list of buckets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BucketResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "storage"
        ],
        "summary": "create a new storage bucket",
        "operationId": "create_bucket",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBucketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "bucket created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BucketResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/buckets/{id}": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "get a single bucket by id",
        "operationId": "get_bucket",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "bucket id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "bucket details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BucketResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "storage"
        ],
        "summary": "delete a storage bucket",
        "operationId": "delete_bucket",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "bucket id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "bucket deleted"
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BucketResponse": {
        "type": "object",
        "description": "bucket response",
        "required": [
          "id",
          "name",
          "access_key",
          "secret_key",
          "endpoint",
          "size_bytes",
          "created_at"
        ],
        "properties": {
          "access_key": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "endpoint": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "secret_key": {
            "type": "string",
            "description": "masked secret key"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "CreateBucketRequest": {
        "type": "object",
        "description": "bucket creation request",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "bucket name"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "error response",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "error message"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}